cleanup links in description👉
I created a Cleanup Dashboard as part of an assignment with Meru Technosoft. The dashboard features six cards, and clicking on each card displays a table with rows and columns of data. Each row includes a dropdown icon under the "Fix Action" column. When clicked, it provides options for Inline Edit and Bulk Edit. Based on the selected option, the remaining rows are set to the corresponding fix action. A search bar is also included, allowing users to filter data by date, description, and account type. Additionally, a progress circle is displayed to indicate progress status.

figma link 
https://www.figma.com/proto/YIJMjy7KWK0H4xjyAwj04q/Untitled?node-id=1-2&p=f&t=JISGxLT7vd2rgdfC-0&scaling=scale-down&content-scaling=fixed&page-id=0%3A1&starting-point-node-id=1%3A2